Clean Up
Cleaned up main files from irelevent modules and comments. Removed .mp4 in gitignore. Made experiment folder. Removed poor quality test data.
